pyproject.toml
[project]
name = "node-copilot"
version = "0.0.1"
description = "An LLM-based Text-to-Code CoPilot for Blender Node Editor"
authors = [
    {name = "Mohammad Wasil Saleem"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "datasets (>=4.1.1,<5.0.0)",
    "torch (==2.8.0)",
    "transformers (>=4.56.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
